Fixed minor logic inverson in MultiLine/PPI.pm
The result was that incomplete q/ / operators
would not be detected and result in an error.
Now they properly continue however, it appears
that q followed by a \n is treaed as a q//
operator.  I don't know if perl allows that
or if this need to be resolved in PPI...